// src/app/interface.tsx
export interface Wgcacak {
    LT_PV_Kg_LT_101N: number;
    LT_PV_Kg_LT_102H: number;
    LT_PV_Kg_LT_102N: number;
    LT_PV_Kg_LT_301H: number;
    LT_PV_Kg_LT_301N: number;
    LT_PV_Kg_LT_401H: number;
    LT_PV_Kg_LT_401N: number;
    LT_PV_m3_LT_101H: number;
    LT_PV_m3_LT_101N: number;
    LT_PV_Kg_LT_101H: number;
    LT_PV_m3_LT_102N: number;
    LT_PV_m3_LT_102H: number;
    LT_PV_m3_LT_301H: number;
    LT_PV_m3_LT_301N: number;
    LT_PV_m3_LT_401H: number;
    LT_PV_m3_LT_401N: number;
    AAA_LT_101H_Output: number;
    AAA_LT_101N_Output: number;
    AAA_LT_102H_Output: number;
    AAA_LT_102N_Output: number;
    AAA_LT_301H_Output: number;
    AAA_LT_301N_Output: number;
    AAA_LT_401H_Output: number;
    AAA_LT_401N_Output: number;
    Auto_Mode_H: number;
    Auto_Mode_N: number;
    PB_Start_Auto_H: number;
    PB_Start_Auto_N: number;
    PB_Stop_Auto_H: number;
    PB_Stop_Auto_N: number;
    AAA_PU_401N_Run: number;
    AAA_PU_401H_Run: number;
    AAA_PU_402N_Run: number;
    AAA_PU_402H_Run: number;
    AAA_PU_403N_Run: number;
    AAA_PU_403H_Run: number;
    DI_AKA_PD01: number;
    DI_AKA_PD02: number;
    DI_AKA_PD03: number;
    DI_ACID_PD01: number;
    DI_ACID_PD02: number;
    DI_ACID_PD03: number;
    Tranfer_SP_Total_PD1_H: number;
    Tranfer_SP_Total_PD1_N: number;
    Tranfer_SP_Total_PD2_H: number;
    Tranfer_SP_Total_PD2_N: number;
    Tranfer_SP_Total_PD3_H: number;
    Tranfer_SP_Total_PD3_N: number;
    AAA_FT_401H_Total: number;
    AAA_FT_401N_Total: number;
    AAA_FT_402H_Total: number;
    AAA_FT_402N_Total: number;
    AAA_FT_403H_Total: number;
    AAA_FT_403N_Total: number;
    ACID_Timer_PD01_PV: number;
    ACID_Timer_PD01_SP: number;
    ACID_Timer_PD02_PV: number;
    ACID_Timer_PD02_SP: number;
    ACID_Timer_PD03_PV: number;
    ACID_Timer_PD03_SP: number;
    AKA_Timer_PD01_PV: number;
    AKA_Timer_PD01_SP: number;
    AKA_Timer_PD02_PV: number;
    AKA_Timer_PD02_SP: number;
    AKA_Timer_PD03_PV: number;
    AKA_Timer_PD03_SP: number;
    Rcipe_H_1: number;
    Rcipe_N_1: number;
}